EES-2990 - refactored CreateClaimsPrincipal() methods into common area not specific to AuthorizationHandlers
